+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 - Fix eliminated error message in Init projection handler [\#2250](https://github.com/reimagined/resolve/pull/2250)
 
+**Other changes:**
+
+- Refactor resolve-scripts and runtimes [\#2245](https://github.com/reimagined/resolve/pull/2245)
+
 ## [V0.33.15](https://github.com/reimagined/resolve/tree/V0.33.15) (2022-03-03)
 
 [Full Changelog](https://github.com/reimagined/resolve/compare/V0.33.14...V0.33.15)